update UI admin

+++ ManageProject/src/main/java/vn/iotstar/Repository/IStudentReposiotory.java
@@ -1,6 +1,7 @@
 package vn.iotstar.Repository;
 
 import java.util.List;
+import java.util.Optional;
 
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
@@ -10,7 +11,7 @@ import org.springframework.stereotype.Repository;
 import vn.iotstar.Entity.Student;
 
 @Repository
-public interface IStudentReposiotory extends JpaRepository<Student, Long>{
+public interface IStudentReposiotory extends JpaRepository<Student, Integer>{
 	List<Student> findByNameContaining(String name);
 
 	List<Student> findByIdproject(Integer idproject); // lấy đúng từng ký tự
@@ -19,4 +20,6 @@ public interface IStudentReposiotory extends JpaRepository<Student, Long>{
 	Page<Student> findByNameContaining(String name, Pageable pageable);
 
 	Student findByEmailContaining(String email);
+
+
 }
